chore(*): organise rules into folders by type
