+++ src/rules/file/hasDoxygenHeader.spec.ts
@@ -1,4 +1,4 @@
-import { Severity } from '../types/Severity'
+import { Severity } from '../../types/Severity'
 import { hasDoxygenHeader } from './hasDoxygenHeader'
 
 describe('hasDoxygenHeader', () => {
