+++ src/rules/file/noNestedMacros.spec.ts
@@ -1,4 +1,4 @@
-import { Severity } from '../types/Severity'
+import { Severity } from '../../types/Severity'
 import { noNestedMacros } from './noNestedMacros'
 
 describe('noNestedMacros', () => {
