+++ src/rules/path/index.ts
@@ -1,0 +1,2 @@
+export { lowerCaseFileNames } from './lowerCaseFileNames'
+export { noSpacesInFileNames } from './noSpacesInFileNames'
