+++ src/rules/path/lowerCaseFileNames.spec.ts
@@ -1,4 +1,4 @@
-import { Severity } from '../types/Severity'
+import { Severity } from '../../types/Severity'
 import { lowerCaseFileNames } from './lowerCaseFileNames'
 
 describe('lowerCaseFileNames', () => {
