+++ src/rules/path/noSpacesInFileNames.spec.ts
@@ -1,4 +1,4 @@
-import { Severity } from '../types/Severity'
+import { Severity } from '../../types/Severity'
 import { noSpacesInFileNames } from './noSpacesInFileNames'
 
 describe('noSpacesInFileNames', () => {
